Move UncVar to separate file
